allow setting required debci exit status
SALSA_CI_AUTOPKGTEST_ALLOWED_EXIT_STATUS contains a list of all exit
status numbers that let the autopkgtest job succeed. By default, this
list contains 0 (all tests passed), 2 (at least one test was skipped or
marked as flaky and failed) and 8 (the package has no tests or all
not-superficial tests were skipped). Individual exit status numbers are
separated by a comma, so the default setting is:

    SALSA_CI_AUTOPKGTEST_ALLOWED_EXIT_STATUS: "0,2,8"

By changing this option, it is possible to be stricter than the default.
For example it is possible to enforce that all tests pass by setting
SALSA_CI_AUTOPKGTEST_ALLOWED_EXIT_STATUS to be equal to "0".