salsa-ci: create /dev/null in piuparts chroot
Without a `/dev/null` special file in the chroot, `apt-key` fails with
the following error:

  /usr/bin/apt-key: 95: cannot create /dev/null: Permission denied

In most cases the job still runs fine as the package index is present in
the image. However, when using an extra repository, the job fails as it
then can't verify the signature for said extra repo.

Manually creating this file fixes this problem.

Fixes #329