Disable failing child job salsa.debian.org/carnil/boxes
The Salsa-CI at salsa.debian.org/carnil/boxes is broken and will not pass.
Thus is is currently unsuitable for regression testing of changes in
Salsa-CI itself. Disable job temporarily by commenting it out. This is
urgent to do as otherwise all Merge Requests that are opened or rebased
will fail and the communication function of CI becomes moot.